code/requirements.txt
torch
numpy
tqdm
spectral
scikit-learn
seaborn
imageio
scikit-image
Pillow
optuna